# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DOT Affiliates - FEARLESS WEEK 2.0 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type AffiliateRegistrationPrimaryPlatform =
  (typeof AffiliateRegistrationPrimaryPlatform)[keyof typeof AffiliateRegistrationPrimaryPlatform];

export const AffiliateRegistrationPrimaryPlatform = {
  instagram: "instagram",
  tiktok: "tiktok",
  twitter: "twitter",
  snapchat: "snapchat",
  whatsapp: "whatsapp",
  facebook: "facebook",
} as const;

export type AffiliateRegistrationTicketsSellEstimate =
  (typeof AffiliateRegistrationTicketsSellEstimate)[keyof typeof AffiliateRegistrationTicketsSellEstimate];

export const AffiliateRegistrationTicketsSellEstimate = {
  "0-50": "0-50",
  "50-100": "50-100",
  "100+": "100+",
} as const;

export interface AffiliateRegistration {
  name: string;
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
  whatsappNumber: string;
  /** @nullable */
  phoneNumber?: string | null;
  primaryPlatform: AffiliateRegistrationPrimaryPlatform;
  avgEngagement: string;
  hasPromotedBefore: boolean;
  whatsappGroupsReach: string;
  ticketsSellEstimate: AffiliateRegistrationTicketsSellEstimate;
  /** @nullable */
  estimatedReach: string | null;
  willingToPromote: boolean;
  whySelectYou: string;
}

export interface LoginInput {
  /** Email or Username */
  identifier: string;
  password: string;
}

export type AuthResponseRole =
  (typeof AuthResponseRole)[keyof typeof AuthResponseRole];

export const AuthResponseRole = {
  admin: "admin",
  affiliate: "affiliate",
} as const;

export type AffiliateStatus =
  (typeof AffiliateStatus)[keyof typeof AffiliateStatus];

export const AffiliateStatus = {
  pending: "pending",
  active: "active",
  suspended: "suspended",
} as const;

export interface Affiliate {
  id: number;
  name: string;
  username: string;
  email: string;
  whatsappNumber?: string;
  /** @nullable */
  phoneNumber?: string | null;
  affiliateCode: string;
  status: AffiliateStatus;
  primaryPlatform: string;
  avgEngagement?: string;
  hasPromotedBefore?: boolean;
  whatsappGroupsReach?: string;
  ticketsSellEstimate?: string;
  /** @nullable */
  estimatedReach?: string | null;
  willingToPromote?: boolean;
  whySelectYou?: string;
  createdAt: string;
}

export interface AuthResponse {
  affiliate?: Affiliate;
  token: string;
  role?: AuthResponseRole;
}

export interface AffiliateStats {
  clicks: number;
  conversions: number;
  /** @nullable */
  rank: number | null;
  affiliateLink: string;
}

export interface AffiliateDashboard {
  affiliate: Affiliate;
  stats: AffiliateStats;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  primaryPlatform: string;
  conversions: number;
  clicks: number;
}

export type AffiliateSettingsUpdatePrimaryPlatform =
  (typeof AffiliateSettingsUpdatePrimaryPlatform)[keyof typeof AffiliateSettingsUpdatePrimaryPlatform];

export const AffiliateSettingsUpdatePrimaryPlatform = {
  instagram: "instagram",
  tiktok: "tiktok",
  twitter: "twitter",
  snapchat: "snapchat",
  whatsapp: "whatsapp",
  facebook: "facebook",
} as const;

export interface AffiliateSettingsUpdate {
  name?: string;
  username?: string;
  whatsappNumber?: string;
  /** @nullable */
  phoneNumber?: string | null;
  primaryPlatform?: AffiliateSettingsUpdatePrimaryPlatform;
}

export interface PasswordChangeInput {
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export type AffiliateDetailStatus =
  (typeof AffiliateDetailStatus)[keyof typeof AffiliateDetailStatus];

export const AffiliateDetailStatus = {
  pending: "pending",
  active: "active",
  suspended: "suspended",
} as const;

export interface AffiliateDetail {
  id: number;
  name: string;
  username: string;
  email: string;
  whatsappNumber?: string;
  /** @nullable */
  phoneNumber?: string | null;
  affiliateCode: string;
  status: AffiliateDetailStatus;
  primaryPlatform: string;
  avgEngagement?: string;
  hasPromotedBefore?: boolean;
  whatsappGroupsReach?: string;
  ticketsSellEstimate?: string;
  /** @nullable */
  estimatedReach?: string | null;
  willingToPromote?: boolean;
  whySelectYou?: string;
  clicks: number;
  conversions: number;
  /** @nullable */
  rank?: number | null;
  createdAt: string;
}

export interface PaginatedAffiliates {
  data: AffiliateDetail[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminStats {
  totalAffiliates: number;
  activeAffiliates: number;
  pendingAffiliates: number;
  suspendedAffiliates: number;
  totalClicks: number;
  totalConversions: number;
  conversionRate: number;
}

export interface ActivityEntry {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  affiliateName?: string | null;
  createdAt: string;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  message: string;
  /** @nullable */
  affiliateId?: number | null;
  /** @nullable */
  affiliateName?: string | null;
  /** @nullable */
  affiliateWhatsapp?: string | null;
  /** @nullable */
  affiliateEmail?: string | null;
  /** @nullable */
  whatsappMessage?: string | null;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationsResponse {
  notifications: Notification[];
  unreadCount: number;
}

export type AdminListAffiliatesParams = {
  status?: AdminListAffiliatesStatus;
  search?: string;
  page?: number;
  limit?: number;
};

export type AdminListAffiliatesStatus =
  (typeof AdminListAffiliatesStatus)[keyof typeof AdminListAffiliatesStatus];

export const AdminListAffiliatesStatus = {
  active: "active",
  suspended: "suspended",
  pending: "pending",
} as const;