# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DOT Affiliates - FEARLESS WEEK 2.0 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/affiliates:
    get:
      operationId: adminListAffiliates
      tags: [admin]
      summary: List all affiliates (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, pending]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Affiliates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAffiliates"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/affiliates/{id}:
    get:
      operationId: adminGetAffiliate
      tags: [admin]
      summary: Get affiliate details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Affiliate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffiliateDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteAffiliate
      tags: [admin]
      summary: Delete affiliate (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/affiliates/{id}/suspend:
    post:
      operationId: adminSuspendAffiliate
      tags: [admin]
      summary: Suspend affiliate (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/affiliates/{id}/unsuspend:
    post:
      operationId: adminUnsuspendAffiliate
      tags: [admin]
      summary: Unsuspend affiliate (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unsuspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/affiliates/{id}/approve:
    post:
      operationId: adminApproveAffiliate
      tags: [admin]
      summary: Approve affiliate (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get admin overview stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/notifications/{id}/read:
    post:
      operationId: adminMarkNotificationRead
      tags: [admin]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    PaginatedAffiliates:
      type: object
      required:
        - data
        - total
        - page
        - limit
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AffiliateDetail"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AffiliateDetail:
      type: object
      required:
        - id
        - name
        - username
        - email
        - affiliateCode
        - status
        - primaryPlatform
        - clicks
        - conversions
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        email:
          type: string
        whatsappNumber:
          type: string
        phoneNumber:
          type: ["string", "null"]
        affiliateCode:
          type: string
        status:
          type: string
          enum: [pending, active, suspended]
        primaryPlatform:
          type: string
        avgEngagement:
          type: string
        hasPromotedBefore:
          type: boolean
        whatsappGroupsReach:
          type: string
        ticketsSellEstimate:
          type: string
        estimatedReach:
          type: ["string", "null"]
        willingToPromote:
          type: boolean
        whySelectYou:
          type: string
        clicks:
          type: integer
        conversions:
          type: integer
        rank:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    AdminStats:
      type: object
      required:
        - totalAffiliates
        - activeAffiliates
        - pendingAffiliates
        - suspendedAffiliates
        - totalClicks
        - totalConversions
        - conversionRate
      properties:
        totalAffiliates:
          type: integer
        activeAffiliates:
          type: integer
        pendingAffiliates:
          type: integer
        suspendedAffiliates:
          type: integer
        totalClicks:
          type: integer
        totalConversions:
          type: integer
        conversionRate:
          type: number
...