# lib/api-zod/src/generated/types/affiliateDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DOT Affiliates - FEARLESS WEEK 2.0 API
 * OpenAPI spec version: 0.1.0
 */
import type { AffiliateDetailStatus } from "./affiliateDetailStatus";

export interface AffiliateDetail {
  id: number;
  name: string;
  username: string;
  email: string;
  whatsappNumber?: string;
  /** @nullable */
  phoneNumber?: string | null;
  affiliateCode: string;
  status: AffiliateDetailStatus;
  primaryPlatform: string;
  avgEngagement?: string;
  hasPromotedBefore?: boolean;
  whatsappGroupsReach?: string;
  ticketsSellEstimate?: string;
  /** @nullable */
  estimatedReach?: string | null;
  willingToPromote?: boolean;
  whySelectYou?: string;
  clicks: number;
  conversions: number;
  /** @nullable */
  rank?: number | null;
  createdAt: Date;
}